requirements-dev.txt
pytest
pytest-flask
requests 